1. 상대 토큰이 제대로 저장되지 않는 문제 수정 2. 제어 명령 형식 수정(FCM을 통해 보낼 때 객체로 전달이 가능하도록 변경)
