+++ .idea/deploymentTargetSelector.xml
@@ -3,7 +3,7 @@
   <component name="deploymentTargetSelector">
     <selectionStates>
       <SelectionState runConfigName="app">
-        <option name="selectionMode" value="DIALOG" />
+        <option name="selectionMode" value="DROPDOWN" />
         <DropdownSelection timestamp="2024-12-04T12:56:08.096141500Z">
           <Target type="DEFAULT_BOOT">
             <handle>
