--- app/src/main/java/com/donghaeng/withme/command/controlcommand/brightnesscontrol/BrightnessControl.java
@@ -1,33 +1,0 @@
-package com.donghaeng.withme.command.controlcommand.brightnesscontrol;
-
-import com.donghaeng.withme.command.controlcommand.ControlCommand;
-
-abstract class BrightnessControl extends ControlCommand{
-    private byte currentBrightness;
-    private byte targetBrightness;
-    //public byte memorizedBrightness;
-    protected BrightnessControl(byte controlCommandType, byte currentBrightness, byte targetBrightness){
-        super(controlCommandType);
-        setCurrentBrightness(currentBrightness);
-        setTargetBrightness(targetBrightness);
-    }
-
-    @Override
-    protected final void cancelControl(){
-        // TODO: 취소 시 원래 밝기로 돌아 가는 기능 구현
-    }
-
-    public final byte getCurrentBrightness() {
-        return currentBrightness;
-    }
-    public final byte getTargetBrightness() {
-        return targetBrightness;
-    }
-
-    private void setCurrentBrightness(byte currentBrightness){
-        this.currentBrightness = currentBrightness;
-    }
-    private void setTargetBrightness(byte targetBrightness) {
-        this.targetBrightness = targetBrightness;
-    }
-}
