--- app/src/main/java/com/donghaeng/withme/command/controlcommand/volumecontrol/MediaVolumeDown.java
@@ -1,9 +1,0 @@
-package com.donghaeng.withme.command.controlcommand.volumecontrol;
-
-import com.donghaeng.withme.data.featurelist.ControlCommandList;
-
-public class MediaVolumeDown extends VolumeControl {
-    public MediaVolumeDown(byte currentVolume, byte targetVolume) {
-        super(ControlCommandList.MEDIA_VOLUME_DOWN, currentVolume, targetVolume);
-    }
-}
