--- app/src/main/java/com/donghaeng/withme/command/controlcommand/volumecontrol/MediaVolumeMute.java
@@ -1,9 +1,0 @@
-package com.donghaeng.withme.command.controlcommand.volumecontrol;
-
-import com.donghaeng.withme.data.featurelist.ControlCommandList;
-
-public class MediaVolumeMute extends VolumeControl {
-    public MediaVolumeMute(byte currentVolume, byte targetVolume) {
-        super(ControlCommandList.MEDIA_VOLUME_MUTE, currentVolume, targetVolume);
-    }
-}
