--- app/src/main/java/com/donghaeng/withme/command/controlcommand/volumecontrol/MediaVolumeUnmute.java
@@ -1,9 +1,0 @@
-package com.donghaeng.withme.command.controlcommand.volumecontrol;
-
-import com.donghaeng.withme.data.featurelist.ControlCommandList;
-
-public class MediaVolumeUnmute extends VolumeControl {
-    public MediaVolumeUnmute(byte currentVolume, byte targetVolume) {
-        super(ControlCommandList.MEDIA_VOLUME_UNMUTE, currentVolume, targetVolume);
-    }
-}
