--- app/src/main/java/com/donghaeng/withme/command/controlcommand/volumecontrol/MediaVolumeUp.java
@@ -1,9 +1,0 @@
-package com.donghaeng.withme.command.controlcommand.volumecontrol;
-
-import com.donghaeng.withme.data.featurelist.ControlCommandList;
-
-public class MediaVolumeUp extends VolumeControl {
-    public MediaVolumeUp(byte currentVolume, byte targetVolume) {
-        super(ControlCommandList.MEDIA_VOLUME_UP, currentVolume, targetVolume);
-    }
-}
