+++ app/src/main/java/com/donghaeng/withme/data/command/LightMode.java
@@ -1,0 +1,7 @@
+package com.donghaeng.withme.data.command;
+
+public class LightMode {
+    public static int CURRENT = 0;
+    public static final int SET = 0;
+    public static final int AUTO = 1;
+}
