+++ app/src/main/java/com/donghaeng/withme/data/command/SoundMode.java
@@ -1,0 +1,8 @@
+package com.donghaeng.withme.data.command;
+
+public class SoundMode {
+    public static int CURRENT = 0;
+    public static final int CALL = 0;
+    public static final int NOTIFICATION = 1;
+    public static final int MEDIA = 2;
+}
