+++ app/src/main/java/com/donghaeng/withme/data/command/controlcommand/ControlCommand.java
@@ -1,0 +1,18 @@
+package com.donghaeng.withme.data.command.controlcommand;
+
+public abstract class ControlCommand {
+    private byte controlCommandType;
+    
+    protected ControlCommand(byte controlCommandType) {
+        setControlCommandType(controlCommandType);
+    }
+
+    protected abstract void cancelControl();
+
+    public final byte getControlCommandType() {
+        return controlCommandType;
+    }
+    private void setControlCommandType(byte controlCommandType) {
+        this.controlCommandType = controlCommandType;
+    }
+}
