+++ app/src/main/java/com/donghaeng/withme/data/command/controlcommand/volumecontrol/CallNormal.java
@@ -1,0 +1,9 @@
+package com.donghaeng.withme.data.command.controlcommand.volumecontrol;
+
+import com.donghaeng.withme.data.featurelist.ControlCommandList;
+
+public class CallNormal extends VolumeControl {
+    public CallNormal(byte currentVolume, byte targetVolume) {
+        super(ControlCommandList.CALL_NORMAL, currentVolume, targetVolume);
+    }
+}
