+++ app/src/main/java/com/donghaeng/withme/data/command/controlcommand/volumecontrol/CallSilent.java
@@ -1,0 +1,9 @@
+package com.donghaeng.withme.data.command.controlcommand.volumecontrol;
+
+import com.donghaeng.withme.data.featurelist.ControlCommandList;
+
+public class CallSilent extends VolumeControl {
+    public CallSilent(byte currentVolume, byte targetVolume) {
+        super(ControlCommandList.CALL_SILENT, currentVolume, targetVolume);
+    }
+}
