+++ app/src/main/java/com/donghaeng/withme/data/command/controlcommand/volumecontrol/CallVibrate.java
@@ -1,0 +1,9 @@
+package com.donghaeng.withme.data.command.controlcommand.volumecontrol;
+
+import com.donghaeng.withme.data.featurelist.ControlCommandList;
+
+public class CallVibrate extends VolumeControl{
+    public CallVibrate(byte currentVolume, byte targetVolume) {
+        super(ControlCommandList.CALL_VIBRATE, currentVolume, targetVolume);
+    }
+}
