+++ app/src/main/java/com/donghaeng/withme/data/command/controlcommand/volumecontrol/CallVolumeDown.java
@@ -1,0 +1,9 @@
+package com.donghaeng.withme.data.command.controlcommand.volumecontrol;
+
+import com.donghaeng.withme.data.featurelist.ControlCommandList;
+
+public class CallVolumeDown extends VolumeControl {
+    public CallVolumeDown(byte currentVolume, byte targetVolume) {
+        super(ControlCommandList.CALL_VOLUME_DOWN, currentVolume, targetVolume);
+    }
+}
