+++ app/src/main/java/com/donghaeng/withme/data/command/controlcommand/volumecontrol/CallVolumeUp.java
@@ -1,0 +1,10 @@
+package com.donghaeng.withme.data.command.controlcommand.volumecontrol;
+
+import com.donghaeng.withme.data.featurelist.ControlCommandList;
+
+public class CallVolumeUp extends VolumeControl {
+    public CallVolumeUp(byte currentVolume, byte targetVolume) {
+        super(ControlCommandList.CALL_VOLUME_UP, currentVolume, targetVolume);
+    }
+
+}
