+++ app/src/main/java/com/donghaeng/withme/data/database/firestore/SendDataMessage.java
@@ -12,6 +12,9 @@ import android.util.Log;
 
 import androidx.annotation.NonNull;
 
+import com.donghaeng.withme.data.message.firebasemessage.FirebaseCloudMessage;
+import com.google.gson.Gson;
+
 import java.io.IOException;
 
 public class SendDataMessage {
@@ -28,6 +31,39 @@ public class SendDataMessage {
                 + "}";
 
         // 요청 만들기a
+        RequestBody requestBody = RequestBody.create(MediaType.parse("application/json"), json);
+
+        Request request = new Request.Builder()
+                .url("https://senddatamessage-2qnis2bswa-uc.a.run.app")  // 실제 URL로 변경
+                .post(requestBody)
+                .build();
+
+        // HTTP 요청 보내기
+        client.newCall(request).enqueue(new Callback() {
+            @Override
+            public void onFailure(@NonNull Call call, @NonNull IOException e) {
+                Log.e("FCM", "Failed to send data message", e);
+            }
+
+            @Override
+            public void onResponse(@NonNull Call call, @NonNull Response response) throws IOException {
+                if (response.isSuccessful()) {
+                    Log.d("FCM", "Data message sent successfully!");
+                    // TODO: 로그 기록 구현
+                } else {
+                    Log.e("FCM", "Error: " + response.code() + ", Body: " + response.body().string());
+                }
+            }
+        });
+    }
+
+    public void sendDataMessage(String token, String type, Object command) {
+        OkHttpClient client = new OkHttpClient();
+
+        FirebaseCloudMessage message = new FirebaseCloudMessage(token, type, command);
+        String json = new Gson().toJson(message);
+
+        // 요청 만들기
         RequestBody requestBody = RequestBody.create(MediaType.parse("application/json"), json);
 
         Request request = new Request.Builder()
