+++ app/src/main/java/com/donghaeng/withme/data/message/nearbymessage/NearbyMessage.java
@@ -1,0 +1,19 @@
+package com.donghaeng.withme.data.message.nearbymessage;
+
+public class NearbyMessage {
+    private final String type;
+    private final Object payload;
+
+    public NearbyMessage(String type, Object payload) {
+        this.type = type;
+        this.payload = payload;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public Object getPayload() {
+        return payload;
+    }
+}
