+++ app/src/main/java/com/donghaeng/withme/data/message/nearbymessage/TextMessagePayload.java
@@ -1,0 +1,13 @@
+package com.donghaeng.withme.data.message.nearbymessage;
+
+public class TextMessagePayload {
+    private final String message;
+
+    public TextMessagePayload(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
